requirements.txt
Hardver (spoj)

Arduino D8 → optokapler (6N137), primarna strana (uz serijski R i pull-up po šemi),

sekundarna strana opta → CCS/H-bridge drajver (TTL ulazi: EN, PHASE),

izlaz CCS na R||C opterećenje ili elektrode (po bezbednosnim metama iz rada),

zajednički povrati i ekranisanje, RC snubber prema potrebi.

Bezbednost

Ove skripte i firmware su za laboratorijsku upotrebu. Za biomedicinske aplikacije obavezna je izolacija i merenje prema IEC 60601-1/-1-2. Radite na sopstvenu odgovornost.

numpy>=1.26
pandas>=2.2
scipy>=1.12
matplotlib>=3.8
pyserial>=3.5
pyyaml>=6.0
